+++ src/index.ts
@@ -11,7 +11,6 @@ import { startCallDispatch } from "./callDispatch.js";
 import { startJoinReminder } from "./joinReminder.js";
 import { startRoleplayHints } from "./roleplayHints.js";
 import { startModCallDetector } from "./modCallDetector.js";
-import { warmUpSttServer } from "./voice/sttServer.js";
 import { warmUpTtsServer } from "./voice/ttsServer.js";
 import { initDb } from "./db.js";
 import { registerInternalApi } from "./internalApi.js";
@@ -85,7 +84,6 @@ async function main() {
   startJoinReminder();
   startRoleplayHints();
   startModCallDetector();
-  warmUpSttServer();
   warmUpTtsServer();
 
   app.listen(PORT, () => {